# src/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Project 1 API",
    "description": "Send forward your gaming community using Project 1 API",
    "version": "1.0.0"
  },
  "produces": [
    "application/json"
  ],
...
  "basePath": "/api",
  "paths": {
...
    "/users": {
      "get": {
        "x-swagger-router-controller": "UserController",
        "operationId": "getUsers",
        "tags": [
          "User"
        ],
        "summary": "Get all users",
        "description": "Gets `User` objects.\nOptional query param of **size** determines\nsize of returned array\n",
        "parameters": [
          {
            "name": "size",
            "in": "query",
            "description": "Size of array",
            "required": false,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of Users",
            "schema": {
              "title": "ArrayOfUsers",
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/users/addUser": {
      "post": {
        "x-swagger-router-controller": "UserController",
        "operationId": "addUser",
        "tags": [
          "User"
        ],
        "summary": "Adds a new user",
        "description": "Adds new `User` object to the DB.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User object to insert",
            "required": false,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User has been successfully added to the DB"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{userId}/getUserById": {
      "get": {
        "x-swagger-router-controller": "UserController",
        "operationId": "getUserById",
        "tags": [
          "User"
        ],
        "summary": "Gets the specified user",
        "description": "Gets `User` object specified by its ID.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User ID to search for.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A 'User' object.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{username}/getUserByUsername": {
      "get": {
        "x-swagger-router-controller": "UserController",
        "operationId": "getUserByUsername",
        "tags": [
          "User"
        ],
        "summary": "Gets the corresponding matching user",
        "description": "Gets `User` object specified by its username.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of User to search for.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A 'User' object.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{userId}/updateUser": {
      "put": {
        "x-swagger-router-controller": "UserController",
        "operationId": "updateUser",
        "tags": [
          "User"
        ],
        "summary": "Updates the corresponding user basic properties",
        "description": "Update `User` object specified by its ID.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId of User object to update.",
            "required": true,
            "type": "string"
          },
          {
            "name": "userObject",
            "in": "body",
            "description": "User object that will update the old one.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A 'User' object.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{userId}/updatePassword": {
      "put": {
        "x-swagger-router-controller": "UserController",
        "operationId": "updatePassword",
        "tags": [
          "User"
        ],
        "summary": "Updates the corresponding user password",
        "description": "Update password attr of `User` object specified by its ID.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId of User object to update.",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "body",
            "description": "Object containing old password (to verify before updating) and new one to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ChangePassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A 'User' object.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{userId}/updateEmail": {
      "put": {
        "x-swagger-router-controller": "UserController",
        "operationId": "updateEmail",
        "tags": [
          "User"
        ],
        "summary": "Updates the corresponding user email",
        "description": "Update email attr of `User` object specified by its ID.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId of User object to update.",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "body",
            "description": "New email to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated User object.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{userId}/deleteUser": {
      "put": {
        "x-swagger-router-controller": "UserController",
        "operationId": "deleteUser",
        "tags": [
          "User"
        ],
        "summary": "Deactivates the corresponding user account",
        "description": "Deactivate `User` object specified by its ID.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId of User object to deactivate.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated User object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/user/verify/{email}": {
      "get": {
        "x-swagger-router-controller": "UserController",
        "operationId": "verifyUserEmail",
        "tags": [
          "Security"
        ],
        "summary": "Verify user email",
        "description": "Verify user email",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email to verify",
            "required": true,
            "type": "string"
          },
          {
            "name": "t",
            "in": "query",
            "description": "token to verify for this email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User email successfully verified"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/register/step0/{email}": {
      "get": {
        "x-swagger-router-controller": "UserController",
        "operationId": "registerUserEmail",
        "tags": [
          "Registration"
        ],
        "summary": "Register user based on his email (step 0)",
        "description": "Register an user based on his email and token and redirect to corresponding registration step",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email to verify and register",
            "required": true,
            "type": "string"
          },
          {
            "name": "t",
            "in": "query",
            "description": "Token to verify this email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User email successfully verified"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "birthDate": {
          "type": "string",
          "format": "date"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "format": "password"
        },
        "avatar": {
          "type": "string"
        },
        "address": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "phoneNumber": {
          "type": "string"
        },
        "admin": {
          "type": "boolean"
        },
        "friends": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        },
        "active": {
          "type": "boolean"
        },
        "created_at": {
          "type": "string",
          "format": "date"
        },
        "updated_at": {
          "type": "string",
          "format": "date"
        }
      }
    },
...
    "Address": {
      "type": "object",
      "properties": {
        "postCode": {
          "type": "integer",
          "format": "int32"
        },
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "line": {
          "type": "string"
        }
      }
    },
...
    "ChangePassword": {
      "type": "object",
      "properties": {
        "oldPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        },
        "newPasswordConfirmation": {
          "type": "string"
        }
      }
    },
...
    "Email": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "email"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    }
  }
}